Add CV style file
I copied this file from using the link on
http://rpi.edu/dept/arc/training/latex/resumes/. The file itself says
it was written by Michael DeCorte in 1989; this style file is apparently
the one everyone uses for their LaTeX CV.

Thanks Michael!

Closes #6.

UPDATE:
It looks like this exact file is hosted on CTAN; I'm not sure why I
didn't find it earlier.

http://www.ctan.org/pkg/res